+++ tracker/tracker/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 17.0.0
 
-- feature flags and user testing functionality is REMOVED.
+- use `InlineCssMode.Inline` by default if testing in localhost and option is not set explicitly (to preserve css files)
+- feature flags and user testing functionality is *REMOVED*
 - support for Long Animation Task API tracking
 - support for timing information in resource tracking
 - Ionic and Web Component compatibility improvements: